move retry logic to object under filesystem
